smtp-mailer/requirements.txt
requests==2.25.1
PyJWT==2.1.0